+++ example/lib/main.dart
@@ -30,7 +30,8 @@ void main() {
     // assetLoader: HttpAssetLoader()
     // assetLoader: FileAssetLoader()
     assetLoader: CsvAssetLoader()
-    // assetLoader: YamlAssetLoader()
+    // assetLoader: YamlAssetLoader() //multiple files
+    // assetLoader: YamlSingleAssetLoader() //single file
     
     // assetLoader: CodegenLoader() 
   ));